+++ src/main/java/model/constraints/GeneralConstraints.java
@@ -30,7 +30,7 @@ public class GeneralConstraints {
                         continue;
                     for (int d = 0; d < parameters.getServices().get(s).getTrafficFlow().getTrafficDemands().size(); d++)
                         expr.addTerm((double) parameters.getServices().get(s).getTrafficFlow().getTrafficDemands().get(d)
-                                / (double) parameters.getLinks().get(l).getAttribute("capacity"), variables.tSPD[s][p][d]);
+                                / (int) parameters.getLinks().get(l).getAttribute("capacity"), variables.tSPD[s][p][d]);
                 }
             if (isMigration)
                 for (int s = 0; s < parameters.getServices().size(); s++)
